requirements.txt
selenium
chromedriver-binary-auto
python-decouple
gcsa
flask
gunicorn
pebble